# specs/006-registration-requests-approval/contracts/registration-requests.openapi.yaml
openapi: 3.1.0
info:
  title: New Talents Registration Requests API
  version: 0.1.0
  description: Feature 006 only. Ordinary JSON never includes document bytes or raw private identity values.
  license: {name: Proprietary, identifier: LicenseRef-Proprietary}
servers:
  - url: /api
...
paths:
...
  /admin/registration-requests:
    get:
      tags: [Administration]
      operationId: listRegistrationRequestsForReview
      summary: List all request types for Administrator review
      parameters:
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/PageSize'}
        - {name: type, in: query, schema: {$ref: '#/components/schemas/RegistrationRequestType'}}
        - {name: status, in: query, schema: {$ref: '#/components/schemas/RegistrationRequestStatus'}}
      responses: {'200': {$ref: '#/components/responses/RequestPage'}, '403': {$ref: '#/components/responses/SafeDenied'}}
  /admin/registration-requests/{requestId}:
    get:
      tags: [Administration]
      operationId: getRegistrationRequestForReview
      summary: Read a restricted Administrator review detail
      parameters: [{$ref: '#/components/parameters/RequestId'}]
      responses: {'200': {description: Restricted review projection, content: {application/json: {schema: {$ref: '#/components/schemas/AdminRequestEnvelope'}}}}, '404': {$ref: '#/components/responses/SafeNotFound'}}
...
  /admin/registration-requests/{requestId}/correction:
    post:
      tags: [Administration]
      operationId: requestRegistrationCorrection
      summary: Return a submitted request for correction
      parameters: [{$ref: '#/components/parameters/RequestId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CorrectionCommand'}}}}
      responses: {'200': {$ref: '#/components/responses/RequestRead'}, '409': {$ref: '#/components/responses/VersionConflict'}}
  /admin/registration-requests/{requestId}/approve:
    post:
      tags: [Administration]
      operationId: approveRegistrationRequest
      summary: Prepare and finalize an approved request safely
      parameters: [{$ref: '#/components/parameters/RequestId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ApprovalCommand'}}}}
      responses: {'202': {description: Approval prepared; evidence deletion/finalization may be pending, content: {application/json: {schema: {$ref: '#/components/schemas/DecisionEnvelope'}}}}, '409': {$ref: '#/components/responses/VersionConflict'}, '422': {$ref: '#/components/responses/ValidationError'}}
  /admin/registration-requests/{requestId}/reject:
    post:
      tags: [Administration]
      operationId: rejectRegistrationRequest
      summary: Reject a request and start evidence deletion
      parameters: [{$ref: '#/components/parameters/RequestId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RejectionCommand'}}}}
      responses: {'200': {description: Rejected and deletion initiated, content: {application/json: {schema: {$ref: '#/components/schemas/DecisionEnvelope'}}}}, '409': {$ref: '#/components/responses/VersionConflict'}}
  /admin/registration-requests/{requestId}/deletion/retry:
    post:
      tags: [Administration]
      operationId: retryRegistrationEvidenceDeletion
      summary: Retry a recoverable evidence deletion
      parameters: [{$ref: '#/components/parameters/RequestId'}]
      requestBody: {$ref: '#/components/requestBodies/Transition'}
      responses: {'202': {description: Retry accepted, content: {application/json: {schema: {$ref: '#/components/schemas/DeletionEnvelope'}}}}, '409': {$ref: '#/components/responses/VersionConflict'}}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    RequestId: {name: requestId, in: path, required: true, schema: {type: string, format: uuid}}
...
    Cursor: {name: cursor, in: query, schema: {type: string, maxLength: 512}}
    PageSize: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
  requestBodies:
    Transition:
      required: true
      content: {application/json: {schema: {$ref: '#/components/schemas/TransitionCommand'}}}
  responses:
...
    RequestRead: {description: Authorized request projection, content: {application/json: {schema: {$ref: '#/components/schemas/RequestEnvelope'}}}}
    RequestPage: {description: Cursor page, content: {application/json: {schema: {$ref: '#/components/schemas/RequestPage'}}}}
    SafeDenied: {description: Safely denied, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
    SafeNotFound: {description: Missing or not authorized without existence disclosure, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
...
    VersionConflict: {description: Stale lifecycle version, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
    ValidationError: {description: Safe field/category validation, content: {application/json: {schema: {$ref: '#/components/schemas/ValidationErrorEnvelope'}}}}
...
  schemas:
...
    RegistrationRequestType: {type: string, enum: [PERSONAL_ADULT, REPRESENTED_MINOR, FORMAL_ACADEMY, NATURAL_PERSON_ACADEMY, ADDITIONAL_ACADEMY_ACCOUNT, ACADEMY_ADULT_PLAYER, ACADEMY_MINOR_PLAYER]}
    RegistrationRequestStatus: {type: string, enum: [DRAFT, SUBMITTED, REQUIRES_CORRECTION, APPROVED, REJECTED]}
    EvidenceStatus: {type: string, enum: [QUARANTINED, SCANNING, CLEAN, REJECTED, REPLACED, DELETION_PENDING, DELETED]}
    DeletionStatus: {type: string, enum: [PENDING, IN_PROGRESS, COMPLETED, RECOVERY_REQUIRED]}
    EvidenceCategory: {type: string, enum: [IDENTITY_FRONT, IDENTITY_BACK, MINOR_CIVIL_IDENTITY, REPRESENTATION_AUTHORITY, RUT, EXISTENCE_CERTIFICATE, RESPONSIBLE_AUTHORITY, OPERATION_PROOF, ADULT_AUTHORIZATION, ACADEMY_ACCOUNT_AUTHORIZATION]}
...
    TransitionCommand: {type: object, additionalProperties: false, required: [expectedVersion, idempotencyKey], properties: {expectedVersion: {type: integer, minimum: 0}, idempotencyKey: {type: string, format: uuid}}}
    CorrectionCommand: {allOf: [{$ref: '#/components/schemas/TransitionCommand'}, {type: object, required: [safeReason, correctionTargets], properties: {safeReason: {type: string, minLength: 1, maxLength: 1000}, correctionTargets: {type: array, minItems: 1, uniqueItems: true, items: {type: string, maxLength: 100}}}}]}
    ApprovalCommand: {allOf: [{$ref: '#/components/schemas/TransitionCommand'}, {type: object, required: [manualDossierConfirmation], properties: {manualDossierConfirmation: {$ref: '#/components/schemas/ManualDossierConfirmation'}}}]}
    RejectionCommand: {allOf: [{$ref: '#/components/schemas/TransitionCommand'}, {type: object, required: [safeReason], properties: {safeReason: {type: string, minLength: 1, maxLength: 1000}}}]}
    ManualDossierConfirmation: {type: object, additionalProperties: false, required: [confirmed, declarationVersion, categories], properties: {confirmed: {type: boolean, const: true}, declarationVersion: {type: string, maxLength: 40}, categories: {type: array, minItems: 1, uniqueItems: true, items: {$ref: '#/components/schemas/EvidenceCategory'}}}}
    EvidenceMetadata: {type: object, additionalProperties: false, required: [id, category, status, sizeBytes], properties: {id: {type: string, format: uuid}, category: {$ref: '#/components/schemas/EvidenceCategory'}, status: {$ref: '#/components/schemas/EvidenceStatus'}, sizeBytes: {type: integer, minimum: 0}, correctionRequired: {type: boolean}}}
    RequestSummary: {type: object, additionalProperties: false, required: [id, type, status, version, capabilities, createdAt], properties: {id: {type: string, format: uuid}, type: {$ref: '#/components/schemas/RegistrationRequestType'}, status: {$ref: '#/components/schemas/RegistrationRequestStatus'}, version: {type: integer, minimum: 0}, safeApplicantLabel: {type: string}, academyLabel: {type: string}, submittedAt: {type: string, format: date-time}, createdAt: {type: string, format: date-time}, evidenceComplete: {type: boolean}, correctionRequired: {type: boolean}, safeReason: {type: string}, capabilities: {type: array, uniqueItems: true, items: {type: string}}}}
    RequestDetail: {allOf: [{$ref: '#/components/schemas/RequestSummary'}, {type: object, properties: {evidence: {type: array, items: {$ref: '#/components/schemas/EvidenceMetadata'}}, deletion: {$ref: '#/components/schemas/DeletionStatusView'}}}]}
    AdminRequestDetail: {allOf: [{$ref: '#/components/schemas/RequestDetail'}, {type: object, properties: {consents: {type: array, items: {$ref: '#/components/schemas/ConsentView'}}, history: {type: array, items: {$ref: '#/components/schemas/RequestEventView'}}, duplicateReview: {$ref: '#/components/schemas/SafeDuplicateReview'}}}]}
    ConsentView: {type: object, additionalProperties: false, required: [type, version, acceptedAt], properties: {type: {type: string}, version: {type: string}, acceptedAt: {type: string, format: date-time}}}
    RequestEventView: {type: object, additionalProperties: false, required: [action, result, at], properties: {action: {type: string}, fromStatus: {$ref: '#/components/schemas/RegistrationRequestStatus'}, toStatus: {$ref: '#/components/schemas/RegistrationRequestStatus'}, result: {type: string}, at: {type: string, format: date-time}}}
    SafeDuplicateReview: {type: object, additionalProperties: false, required: [state], properties: {state: {type: string, enum: [CLEAR, REVIEW_REQUIRED, RESOLVED_DISTINCT, CONFLICT]}, canApprove: {type: boolean}}}
    DeletionStatusView: {type: object, additionalProperties: false, required: [status, totalItems, completedItems], properties: {status: {$ref: '#/components/schemas/DeletionStatus'}, totalItems: {type: integer, minimum: 0}, completedItems: {type: integer, minimum: 0}, lastUpdatedAt: {type: string, format: date-time}}}
    DecisionView: {type: object, additionalProperties: false, required: [requestId, requestStatus, deletion], properties: {requestId: {type: string, format: uuid}, requestStatus: {$ref: '#/components/schemas/RegistrationRequestStatus'}, deletion: {$ref: '#/components/schemas/DeletionStatusView'}}}
    Pagination: {type: object, additionalProperties: false, required: [hasMore], properties: {nextCursor: {type: string}, hasMore: {type: boolean}}}
    RequestEnvelope: {type: object, additionalProperties: false, required: [data], properties: {data: {$ref: '#/components/schemas/RequestDetail'}}}
    AdminRequestEnvelope: {type: object, additionalProperties: false, required: [data], properties: {data: {$ref: '#/components/schemas/AdminRequestDetail'}}}
...
    DeletionEnvelope: {type: object, additionalProperties: false, required: [data], properties: {data: {$ref: '#/components/schemas/DeletionStatusView'}}}
    DecisionEnvelope: {type: object, additionalProperties: false, required: [data], properties: {data: {$ref: '#/components/schemas/DecisionView'}}}
    RequestPage: {type: object, additionalProperties: false, required: [data, pagination], properties: {data: {type: array, items: {$ref: '#/components/schemas/RequestSummary'}}, pagination: {$ref: '#/components/schemas/Pagination'}}}
    ErrorEnvelope: {type: object, additionalProperties: false, required: [error], properties: {error: {type: object, additionalProperties: false, required: [code, message], properties: {code: {type: string}, message: {type: string}, correlationId: {type: string}}}}}
    ValidationIssue: {type: object, additionalProperties: false, required: [field, code], properties: {field: {type: string}, code: {type: string}, message: {type: string}}}
    ValidationErrorEnvelope: {allOf: [{$ref: '#/components/schemas/ErrorEnvelope'}, {type: object, properties: {issues: {type: array, items: {$ref: '#/components/schemas/ValidationIssue'}}}}]}
security:
  - bearerAuth: []